+++ luci-app-pushbot/luasrc/model/cbi/pushbot/client.lua
@@ -1,0 +1,6 @@
+f = SimpleForm("pushbot")
+luci.sys.call("/usr/bin/pushbot/pushbot client")
+f.reset = false
+f.submit = false
+f:append(Template("pushbot/client"))
+return f
