+++ luci-app-pushbot/luasrc/model/cbi/pushbot/log.lua
@@ -1,0 +1,5 @@
+f = SimpleForm("pushbot")
+f.reset = false
+f.submit = false
+f:append(Template("pushbot/log"))
+return f
