+++ luci-app-pushbot/luasrc/view/pushbot/pushbot_status.htm
@@ -1,0 +1,22 @@
+<script type="text/javascript">//<![CDATA[
+XHR.poll(3, '<%=url([[admin]], [[services]], [[pushbot]], [[status]])%>', null,
+	function(x, data) {
+		var tb = document.getElementById('pushbot_status');
+		if (data && tb) {
+			if (data.running) {
+				var links = '<em><b><font color=green>pushbot <%:RUNNING%></font></b></em>';
+				tb.innerHTML = links;
+			} else {
+				tb.innerHTML = '<em><b><font color=red>pushbot <%:NOT RUNNING%></font></b></em>';
+			}
+		}
+	}
+);
+//]]>
+</script>
+<style>.mar-10 {margin-left: 50px; margin-right: 10px;}</style>
+<fieldset class="cbi-section">
+	<p id="pushbot_status">
+		<em><%:Collecting data...%></em>
+	</p>
+</fieldset>
