+++ luci-app-pushbot/root/etc/config/pushbot
@@ -1,0 +1,11 @@
+
+config pushbot 'pushbot'
+	option pushbot_enable '0'
+	option sleeptime '60'
+	option pushbot_ipv6 '0'
+	option pushbot_up '1'
+	option pushbot_down '1'
+	option cpuload_enable '1'
+	option cpuload '2'
+	option temperature_enable '0'
+
