+++ luci-app-pushbot/root/etc/init.d/pushbot
@@ -1,0 +1,26 @@
+#!/bin/sh /etc/rc.common
+
+START=99
+STOP=10
+
+start() {
+	state=`pgrep -f "/usr/bin/pushbot/pushbot"`
+	if [ ! -z "$state" ]; then
+		restart
+	else
+		/usr/bin/pushbot/pushbot &
+	fi
+	echo "pushbot is starting now ..."
+}
+
+stop() {
+	kill -9 `pgrep -f "/usr/bin/pushbot/pushbot"` 2>/dev/null
+	echo "pushbot exit ..."
+}
+
+restart(){
+	stop
+	sleep 1
+	start
+	echo "restarted."
+}
