+++ luci-app-pushbot/root/etc/uci-defaults/luci-pushbot
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+uci -q batch <<-EOF >/dev/null
+	delete ucitrack.@pushbot[-1]
+	add ucitrack pushbot
+	set ucitrack.@pushbot[-1].init=pushbot
+	commit ucitrack
+EOF
+
+rm -rf /tmp/luci-*
+exit 0
