+++ luci-app-pushbot/root/usr/share/rpcd/acl.d/luci-app-pushbot.json
@@ -1,0 +1,11 @@
+{
+	"luci-app-pushbot": {
+		"description": "Grant UCI access for luci-app-pushbot",
+		"read": {
+			"uci": [ "pushbot" ]
+		},
+		"write": {
+			"uci": [ "pushbot" ]
+		}
+	}
+}
