--- luci-app-serverchan/luasrc/model/cbi/serverchan/log.lua
@@ -1,5 +1,0 @@
-f = SimpleForm("serverchan")
-f.reset = false
-f.submit = false
-f:append(Template("serverchan/log"))
-return f
