--- luci-app-serverchan/luasrc/view/serverchan/serverchan_status.htm
@@ -1,22 +1,0 @@
-<script type="text/javascript">//<![CDATA[
-XHR.poll(3, '<%=url([[admin]], [[services]], [[serverchan]], [[status]])%>', null,
-	function(x, data) {
-		var tb = document.getElementById('serverchan_status');
-		if (data && tb) {
-			if (data.running) {
-				var links = '<em><b><font color=green>serverchan <%:RUNNING%></font></b></em>';
-				tb.innerHTML = links;
-			} else {
-				tb.innerHTML = '<em><b><font color=red>serverchan <%:NOT RUNNING%></font></b></em>';
-			}
-		}
-	}
-);
-//]]>
-</script>
-<style>.mar-10 {margin-left: 50px; margin-right: 10px;}</style>
-<fieldset class="cbi-section">
-	<p id="serverchan_status">
-		<em><%:Collecting data...%></em>
-	</p>
-</fieldset>
