--- luci-app-serverchan/root/etc/config/serverchan
@@ -1,11 +1,0 @@
-
-config serverchan 'serverchan'
-	option serverchan_enable '0'
-	option sleeptime '60'
-	option serverchan_ipv6 '0'
-	option serverchan_up '1'
-	option serverchan_down '1'
-	option cpuload_enable '1'
-	option cpuload '2'
-	option temperature_enable '0'
-
