--- luci-app-serverchan/root/etc/init.d/serverchan
@@ -1,26 +1,0 @@
-#!/bin/sh /etc/rc.common
-
-START=99
-STOP=10
-
-start() {
-	state=`pgrep -f "/usr/bin/serverchan/serverchan"`
-	if [ ! -z "$state" ]; then
-		restart
-	else
-		/usr/bin/serverchan/serverchan &
-	fi
-	echo "serverchan is starting now ..."
-}
-
-stop() {
-	kill -9 `pgrep -f "/usr/bin/serverchan/serverchan"` 2>/dev/null
-	echo "serverchan exit ..."
-}
-
-restart(){
-	stop
-	sleep 1
-	start
-	echo "restarted."
-}
