--- luci-app-serverchan/root/etc/uci-defaults/luci-serverchan
@@ -1,11 +1,0 @@
-#!/bin/sh
-
-uci -q batch <<-EOF >/dev/null
-	delete ucitrack.@serverchan[-1]
-	add ucitrack serverchan
-	set ucitrack.@serverchan[-1].init=serverchan
-	commit ucitrack
-EOF
-
-rm -rf /tmp/luci-*
-exit 0
