--- luci-app-serverchan/root/usr/bin/serverchan/api/ip_blacklist
@@ -1,1 +1,0 @@
-create ip_blacklist hash:ip family inet hashsize 1024 maxelem 65536
