--- luci-app-serverchan/root/usr/bin/serverchan/api/ipv4.list
@@ -1,6 +1,0 @@
-www.cip.cc
-ipv4.ddnspod.com
-ifcfg.cn
-speed.neu.edu.cn/getIP.php
-ddns.oray.com/checkip
-www.net.cn/static/customercare/yourip.asp
