--- luci-app-serverchan/root/usr/bin/serverchan/api/ipv6.list
@@ -1,5 +1,0 @@
-ip.sb
-ipv6.ddnspod.com
-api-ipv6.ip.sb/ip
-speed.neu6.edu.cn/getIP.php
-v6.myip.la/json
