--- luci-app-serverchan/root/usr/bin/serverchan/api/pushplus.json
@@ -1,21 +1,0 @@
-{
-    "_api": "这是 pushplus api 文件",
-    "_api": "【pushplus】",
-    
-    "url": "http://www.pushplus.plus/send",
-    "data": "@${tempjsonpath}",
-    "content_type": "Content-Type: application/json",
-    "str_title_start": "##### ",
-    "str_title_end": "",
-    "str_linefeed": "\\n",
-    "str_splitline": "\\n----\\n",
-    "str_space": " ",
-    "str_tab": "    ",
-    "type":
-      {
-        "title": "\"${1}\"",
-        "content": "\"${2}\"",
-        "token": "\"${pushplus_token}\"",
-        "template":"\"markdown\""
-      }
-}
