--- luci-app-serverchan/root/usr/bin/serverchan/api/qywx_markdown.json
@@ -1,23 +1,0 @@
-{
-    "_api": "这是企业微信 markdown 模板信息 api 文件",
-    "_api": "【企业微信】",
-    
-    "url": "https://qyapi.weixin.qq.com/cgi-bin/message/send?access_token=$(curl -s \"https://qyapi.weixin.qq.com/cgi-bin/gettoken?corpid=${corpid}&corpsecret=${corpsecret}\"|/usr/bin/jq '.access_token' | sed 's/\"//g')",
-    "data": "@${tempjsonpath}",
-    "content_type": "Content-Type: application/json",
-    "str_title_start": "#### ",
-    "str_title_end": "",
-    "str_linefeed": "\\n",
-    "str_splitline": "\\n\\n",
-    "str_space": " ",
-    "str_tab": "> ",
-    "type":
-      {
-		"touser": "\"${userid}\"",
-        "msgtype": "\"markdown\"",
-        "agentid": "\"${agentid}\"",
-        "markdown": {
-                    "content": "\"${1}${str_linefeed}${nowtime}${2}\""
-                }
-        }
-}
