--- luci-app-serverchan/root/usr/bin/serverchan/api/qywx_mpnews.json
@@ -1,33 +1,0 @@
-{
-    "_api": "这是企业微信图文信息 api 文件",
-    "_api": "【企业微信】",
-    
-    "url": "\"https://qyapi.weixin.qq.com/cgi-bin/message/send?access_token=$(curl -s \"https://qyapi.weixin.qq.com/cgi-bin/gettoken?corpid=${corpid}&corpsecret=${corpsecret}\"|/usr/bin/jq '.access_token'|sed 's/\"//g')\"",
-    "data": "@${tempjsonpath}",
-    "content_type": "Content-Type: application/json",
-    "str_title_start": "<h3>",
-    "str_title_end": "</h3>",
-    "str_linefeed": "\\n",
-    "str_splitline": "<hr>",
-    "str_space": " ",
-    "str_tab": "<li>",
-    "type":
-      {
-		"touser": "\"${userid}\"",
-        "msgtype": "\"mpnews\"",
-        "agentid": "\"${agentid}\"",
-        "mpnews":{
-        "articles":[
-            {
-                "title": "\"${nowtime}${str_linefeed}${1}\"",
-                "thumb_media_id": "\"`curl \"https://qyapi.weixin.qq.com/cgi-bin/media/upload?access_token=$(curl -s \"https://qyapi.weixin.qq.com/cgi-bin/gettoken?corpid=${corpid}&corpsecret=${corpsecret}\"|/usr/bin/jq '.access_token'|sed 's/\"//g')&type=image\" -F \"file=@${mediapath}\"|/usr/bin/jq '.media_id'|sed 's/\"//g'`\"",
-                "author": "\"\"",
-                "content_source_url": "\"\"",
-                "content": "\"${2}\"",
-                "digest": "\"\""
-            }
-         ]
-    },
-    "safe":0
-        }
-}
