--- luci-app-serverchan/root/usr/bin/serverchan/api/serverchan.json
@@ -1,17 +1,0 @@
-{
-    "_api": "这是 serverchan api 文件",
-    "_api": "【serverchan】",
-    
-    "url": "\"https://sctapi.ftqq.com/${sckey}.send\"",
-    "data": "\"text=${1}&desp=${nowtime}${str_linefeed}${2}\"",
-    "content_type": "Content-Type:application/x-www-form-urlencoded",
-    "str_title_start": "#### ",
-    "str_title_end": "",
-    "str_linefeed": "%0D%0A%0D%0A",
-    "str_splitline": "%0D%0A%0D%0A----%0D%0A%0D%0A",
-    "str_space": " ",
-    "str_tab": "    ",
-    "type":
-      {
-      }
-}
