--- luci-app-serverchan/root/usr/bin/serverchan/api/telegram.json
@@ -1,20 +1,0 @@
-{
-    "_api": "这是 telegram api 文件",
-    "_api": "【telegram】",
-    
-    "url": "https://api.telegram.org/bot${tg_token}/sendMessage",
-    "data": "@${tempjsonpath}",
-    "content_type": "Content-Type: application/json",
-    "str_title_start": "<b>",
-    "str_title_end": "</b>",
-    "str_linefeed": "\\n",
-    "str_splitline": "\\n----\\n",
-    "str_space": " ",
-    "str_tab": "    ",
-    "type":
-      {
-		"text":"\"${str_title_start}${1}${str_title_end}${str_splitline}${nowtime}${2}\"",
-		"chat_id":"\"${chat_id}\"",
-        "parse_mode":"\"HTML\""
-      }
-}
