--- luci-app-serverchan/root/usr/bin/serverchan/api/wxpusher.json
@@ -1,23 +1,0 @@
-{
-    "_api": "这是 wxpusher api 文件",
-    "_api": "【wxpusher】",
-    
-    "url": "http://wxpusher.zjiecode.com/api/send/message",
-    "data": "@${tempjsonpath}",
-    "content_type": "Content-Type: application/json",
-    "str_title_start": "#### ",
-    "str_title_end": "",
-    "str_linefeed": "\\n",
-    "str_splitline": "\\n----\\n",
-    "str_space": " ",
-    "str_tab": "    ",
-    "type":
-      {
-		"summary":"\"${1}\"",
-		"content":"\"${2}\"",
-		"appToken":"\"${wxpusher_apptoken}\"",
-        "topicIds":"[\"${wxpusher_topicIds}\"]",
-        "uids":"[\"${wxpusher_uids}\"]",
-        "contentType":3
-      }
-}
