--- luci-app-serverchan/root/usr/share/rpcd/acl.d/luci-app-serverchan.json
@@ -1,11 +1,0 @@
-{
-	"luci-app-serverchan": {
-		"description": "Grant UCI access for luci-app-serverchan",
-		"read": {
-			"uci": [ "serverchan" ]
-		},
-		"write": {
-			"uci": [ "serverchan" ]
-		}
-	}
-}
